Allow password-less proxySettings
It turns out that in some enterprise customer the proxy auth settings include a
username, but not a password: microsoft/AzureStorageExplorer#4430

This change allows password-less auth settings to be passed to the underlying agent.

It's unclear whether it's a real scenario when a proxy auth has a password, but
not username so this PR doesn't consider this scenario.